Moved stats to 'message' module
