+++ common/message/src/main/java/org/thingsboard/server/common/msg/stats/StatsType.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.thingsboard.server.queue.stats;
+package org.thingsboard.server.common.msg.stats;
 
 public enum StatsType {
     RULE_ENGINE("ruleEngine"), CORE("core"), TRANSPORT("transport"), JS_INVOKE("jsInvoke");
