+++ common/queue/src/main/java/org/thingsboard/server/queue/TbQueueRequestTemplate.java
@@ -16,7 +16,7 @@
 package org.thingsboard.server.queue;
 
 import com.google.common.util.concurrent.ListenableFuture;
-import org.thingsboard.server.queue.stats.MessagesStats;
+import org.thingsboard.server.common.msg.stats.MessagesStats;
 
 public interface TbQueueRequestTemplate<Request extends TbQueueMsg, Response extends TbQueueMsg> {
